allow nodes to be given an initial balance
Before, transactions failed since the initial balance of nodes was
always 0. Now, the tracker gives each node a starting balance which is
taken into account when checking the validity of a transaction.

The limitation right now is that all nodes must have the same starting
balance. This starting balance is currently set to 10.